fix по замечаниям 3

+++ controllers/cards.js
@@ -1,6 +1,7 @@
 const Card = require('../models/card');
 const NotFoundError = require('../errors/not-found-err');
 const BadRequestError = require('../errors/bad-reques-err');
+const ForbiddenError = require('../errors/forbidden-err');
 
 module.exports.getCards = (req, res, next) => {
   Card.find({})
@@ -14,14 +15,29 @@ module.exports.createCard = (req, res, next) => {
   const owner = req.user._id;
   Card.create({ name, link, owner })
     .then((card) => res.send(card))
-    .catch(next);
+    .catch((err) => {
+      if (err.name === 'ValidationError') {
+        next(new BadRequestError('Переданы некорректные данные'));
+      } else {
+        next(err);
+      }
+    });
 };
 
 module.exports.deleteCard = (req, res, next) => {
   const owner = req.user._id;
-  Card.deleteOne({ _id: req.params.cardId, owner })
+  Card.findOne({ _id: req.params.cardId })
     .orFail(() => new NotFoundError('Нет карточки с таким id'))
-    .then((item) => res.send(item))
+    .then((card) => {
+      if (card.owner.toString() === owner) {
+        Card.deleteOne({ _id: req.params.cardId, owner })
+          .orFail(() => new ForbiddenError('Отказ в авторизации запроса'))
+          .then((item) => res.send(item))
+          .catch(next);
+      } else {
+        throw new ForbiddenError('Отказ в авторизации запроса');
+      }
+    })
     .catch((err) => {
       if (err.name === 'CastError') {
         next(new BadRequestError('Невалидный id'));
